research/D3VAE/requirements.txt
astor==0.8.1
einops==0.4.1
fonttools==4.31.2
gluonts==0.9.2
hijri-converter==2.2.3
holidays==0.13
llvmlite==0.39.1
lmdb==1.3.0
paddlepaddle-gpu==2.3.2
pandas==1.4.1
Pillow==9.0.1
properscoring==0.1
scipy==1.8.0
seaborn==0.11.2
tensorboardX~=2.5.1
tqdm==4.63.1
tslearn==0.5.2
zipp==3.8.1
